requirements.txt
# Core
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-core>=0.2.0

# 智谱AI
zhipuai>=2.0.0

# RAG & Embeddings
langchain-chroma>=0.1.0
chromadb>=0.4.0

# Document Loaders
pypdf>=3.0.0
python-docx>=0.8.11
beautifulsoup4>=4.12.0
lxml>=4.9.0
requests>=2.31.0

# Web UI
streamlit>=1.30.0

# PDF导出
reportlab>=4.0.0

# Utils
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0

# Testing
pytest>=7.0.0
